+++ src/main/java/de/samply/qualityreportgenerator/context/CsvConfig.java
@@ -1,0 +1,11 @@
+package de.samply.qualityreportgenerator.context;
+
+import java.nio.charset.Charset;
+
+public record CsvConfig(
+    Charset charset,
+    String endOfLine,
+    String delimiter
+) {
+
+}