+++ src/main/java/de/samply/qualityreportgenerator/report/QualityReportGenerator.java
@@ -1,17 +1,19 @@
 package de.samply.qualityreportgenerator.report;
 
 import de.samply.qualityreportgenerator.app.QrgConst;
+import de.samply.qualityreportgenerator.context.Context;
+import de.samply.qualityreportgenerator.context.ContextGenerator;
 import de.samply.qualityreportgenerator.exporter.ExporterClient;
 import de.samply.qualityreportgenerator.exporter.ExporterClientException;
+import de.samply.qualityreportgenerator.script.ScriptEngineManager;
+import de.samply.qualityreportgenerator.script.ScriptResult;
 import de.samply.qualityreportgenerator.template.QualityReportTemplate;
-import de.samply.qualityreportgenerator.template.QualityReportTemplateManager;
 import de.samply.qualityreportgenerator.utils.VariablesReplacer;
 import de.samply.qualityreportgenerator.zip.ExporterUnzipper;
 import de.samply.qualityreportgenerator.zip.ExporterUnzipperException;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.nio.file.Path;
-import java.util.Arrays;
 import org.apache.poi.ss.usermodel.Workbook;
 import org.apache.poi.xssf.streaming.SXSSFWorkbook;
 import org.springframework.beans.factory.annotation.Value;
@@ -23,6 +25,8 @@ public class QualityReportGenerator {
   private final ExporterClient exporterClient;
   private final ExporterUnzipper exporterUnzipper;
   private final VariablesReplacer variablesReplacer;
+  private final ScriptEngineManager scriptEngineManager;
+  private final ContextGenerator contextGenerator;
   private final Integer workbookWindow;
   private final Path qualityReportsDirectory;
 
@@ -31,12 +35,16 @@ public class QualityReportGenerator {
       ExporterClient exporterClient,
       ExporterUnzipper exporterUnzipper,
       VariablesReplacer variablesReplacer,
+      ScriptEngineManager scriptEngineManager,
+      ContextGenerator contextGenerator,
       @Value(QrgConst.EXCEL_WORKBOOK_WINDOW_SV) int workbookWindow,
       @Value(QrgConst.QUALITY_REPORTS_DIRECTORY_SV) String qualityReportsDirectory
   ) {
     this.exporterClient = exporterClient;
     this.exporterUnzipper = exporterUnzipper;
     this.variablesReplacer = variablesReplacer;
+    this.scriptEngineManager = scriptEngineManager;
+    this.contextGenerator = contextGenerator;
     this.workbookWindow = workbookWindow;
     this.qualityReportsDirectory = Path.of(qualityReportsDirectory);
   }
@@ -51,6 +59,8 @@ public class QualityReportGenerator {
 
   private void generate(QualityReportTemplate template, String filePath) {
     Path[] paths = extractPaths(filePath);
+    Context context = contextGenerator.generate(template, paths);
+    ScriptResult[] scriptResults = scriptEngineManager.generateRawQualityReport(template, context);
     // TODO
     Workbook workbook = new SXSSFWorkbook(workbookWindow);
     // TODO