+++ src/main/java/de/samply/qualityreportgenerator/script/ScriptEngine.java
@@ -1,0 +1,12 @@
+package de.samply.qualityreportgenerator.script;
+
+import de.samply.qualityreportgenerator.context.Context;
+import de.samply.qualityreportgenerator.template.script.Script;
+import de.samply.qualityreportgenerator.template.script.ScriptFramework;
+
+public interface ScriptEngine {
+
+  ScriptFramework getScriptFramework();
+  ScriptResult generateRawResult(Script script, Context context) throws ScriptEngineException;
+
+}