+++ src/main/java/de/samply/qualityreportgenerator/script/ScriptResult.java
@@ -1,0 +1,17 @@
+package de.samply.qualityreportgenerator.script;
+
+import java.nio.file.Path;
+
+public class ScriptResult {
+
+  private Path rawResult;
+
+  public Path getRawResult() {
+    return rawResult;
+  }
+
+  public void setRawResult(Path rawResult) {
+    this.rawResult = rawResult;
+  }
+
+}